pyproject.toml
[project]
name = "m2rs-0-1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.2.0",
    "pillow>=12.0.0",
    "torch>=2.9.0",
    "tqdm>=4.67.1",
    "transformers>=4.57.1",
]
